refactor: move game setup functions to systems/game_logic.py
- Move new_star(), spawn_starfield(), and start_game() from main.py to systems/game_logic.py
- Remove redundant imports from main.py
- Add necessary imports (Starfield, Player, etc.) to systems/game_logic.py
- Update calls to use game_logic.start_game(game) in state transitions
- Cleaner separation: main.py focuses on game loop, game_logic handles game setup
